backend/requirements.txt
django==3.2
djangorestframework==3.14.0
djoser==2.2.0
Pillow==10.0.0
psycopg2-binary==2.9.6
drf-extra-fields==3.6.1
